Move actions into resources
