+++ libraries/_autoload.rb
@@ -1,0 +1,2 @@
+$LOAD_PATH.unshift *Dir[File.expand_path('../../files/default/vendor/gems/**/lib', __FILE__)]
+$LOAD_PATH.unshift *Dir[File.expand_path('..', __FILE__)]
