+++ libraries/docker_image.rb
@@ -1,10 +1,10 @@
+require 'docker'
+require 'helpers_image'
+
 class Chef
   class Resource
-    class DockerImage < ChefCompat::Resource
+    class DockerImage < DockerBase
       use_automatic_resource_name
-
-      allowed_actions :build, :build_if_missing, :import, :pull, :pull_if_missing, :push, :remove, :save
-      default_action :pull
 
       # https://docs.docker.com/reference/api/docker_remote_api_v1.20/
       property :api_retries, kind_of: Fixnum, default: 3
@@ -24,6 +24,59 @@ class Chef
       alias_method :image_name, :repo
       alias_method :no_cache, :nocache
       alias_method :no_prune, :noprune
+
+      #########
+      # Actions
+      #########
+
+      default_action :pull
+
+      declare_action_class.class_eval do
+        include DockerHelpers::Image
+      end
+
+      action :build do
+        build_image
+        new_resource.updated_by_last_action(true)
+      end
+
+      action :build_if_missing do
+        next if Docker::Image.exist?(image_identifier, {}, connection)
+        action_build
+        new_resource.updated_by_last_action(true)
+      end
+
+      action :import do
+        next if Docker::Image.exist?(image_identifier, {}, connection)
+        import_image
+        new_resource.updated_by_last_action(true)
+      end
+
+      action :pull do
+        r = pull_image
+        new_resource.updated_by_last_action(r)
+      end
+
+      action :pull_if_missing do
+        next if Docker::Image.exist?(image_identifier, {}, connection)
+        action_pull
+      end
+
+      action :push do
+        push_image
+        new_resource.updated_by_last_action(true)
+      end
+
+      action :remove do
+        next unless Docker::Image.exist?(image_identifier, {}, connection)
+        remove_image
+        new_resource.updated_by_last_action(true)
+      end
+
+      action :save do
+        save_image
+        new_resource.updated_by_last_action(true)
+      end
     end
   end
 end
