+++ libraries/docker_tag.rb
@@ -3,13 +3,25 @@ class Chef
     class DockerTag < ChefCompat::Resource
       use_automatic_resource_name
 
-      allowed_actions :tag
-      default_action :tag
-
       property :target_repo, kind_of: String, name_attribute: true
       property :target_tag, kind_of: String
       property :to_repo, kind_of: String
       property :to_tag, kind_of: String
+
+      #########
+      # Actions
+      #########
+
+      action :tag do
+        next if Docker::Image.exist?("#{new_resource.to_repo}:#{new_resource.to_tag}")
+        begin
+          i = Docker::Image.get("#{new_resource.target_repo}:#{new_resource.target_tag}")
+          i.tag('repo' => new_resource.to_repo, 'tag' => new_resource.to_tag, 'force' => true)
+        rescue Docker::Error => e
+          raise e.message
+        end
+        new_resource.updated_by_last_action(true)
+      end
     end
   end
 end
