--- libraries/provider_docker_image.rb
@@ -1,70 +1,0 @@
-$LOAD_PATH.unshift *Dir[File.expand_path('../../files/default/vendor/gems/**/lib', __FILE__)]
-require 'docker'
-require_relative 'helpers_image'
-require_relative 'helpers_auth'
-require_relative 'helpers_connection'
-
-class Chef
-  class Provider
-    class DockerImage < Chef::Provider::LWRPBase
-      # register with the resource resolution system
-      provides :docker_image if Chef::Provider.respond_to?(:provides)
-
-      include DockerHelpers::Image
-      include DockerHelpers::Authentication
-      include DockerHelpers::Connection
-
-      def initialize(*args)
-        super
-        @conn = Docker::Connection.new(parsed_connect_host, parsed_connect_options)
-      end
-
-      #########
-      # Actions
-      #########
-
-      action :build do
-        build_image
-        new_resource.updated_by_last_action(true)
-      end
-
-      action :build_if_missing do
-        next if Docker::Image.exist?(image_identifier, {}, @conn)
-        action_build
-        new_resource.updated_by_last_action(true)
-      end
-
-      action :import do
-        next if Docker::Image.exist?(image_identifier, {}, @conn)
-        import_image
-        new_resource.updated_by_last_action(true)
-      end
-
-      action :pull do
-        r = pull_image
-        new_resource.updated_by_last_action(r)
-      end
-
-      action :pull_if_missing do
-        next if Docker::Image.exist?(image_identifier, {}, @conn)
-        action_pull
-      end
-
-      action :push do
-        push_image
-        new_resource.updated_by_last_action(true)
-      end
-
-      action :remove do
-        next unless Docker::Image.exist?(image_identifier, {}, @conn)
-        remove_image
-        new_resource.updated_by_last_action(true)
-      end
-
-      action :save do
-        save_image
-        new_resource.updated_by_last_action(true)
-      end
-    end
-  end
-end
