+++ spec/docker_service_test/sysvinit_spec.rb
@@ -1,11 +1,11 @@
 # require 'spec_helper'
 
-# describe 'Chef::Provider::DockerService::Sysvinit with Centos' do
+# describe 'docker_service with Centos' do
 #   let(:chef_run) do
 #     ChefSpec::SoloRunner.new(
 #       platform: 'centos',
 #       version: '7.0',
-#       step_into: 'docker_service'
+#       step_into: 'docker_service_sysvinit'
 #     ).converge('docker_service_test::sysvinit')
 #   end
 
@@ -13,8 +13,8 @@
 #     stub_command('/usr/bin/docker  ps | head -n 1 | grep ^CONTAINER').and_return(true)
 #   end
 
-#   it 'creates docker_service[default]' do
-#     expect(chef_run).to create_docker_service('default')
+#   it 'creates docker_service_sysvinit[default]' do
+#     expect(chef_run).to create_docker_service_sysvinit('default')
 #   end
 
 #   it 'creates the sysvinit file' do
