+++ test/cookbooks/docker_service_test/recipes/tls.rb
@@ -106,7 +106,7 @@ bash 'signing request for client' do
 end
 
 # start docker service listening on TCP port
-docker_service 'tls_test:2376' do
+service_def = proc do
   host 'tcp://127.0.0.1:2376'
   tls_verify true
   tls_ca_cert "#{caroot}/ca.pem"
@@ -114,6 +114,11 @@ docker_service 'tls_test:2376' do
   tls_server_key "#{caroot}/server-key.pem"
   tls_client_cert "#{caroot}/cert.pem"
   tls_client_key "#{caroot}/key.pem"
-  provider Chef::Provider::DockerService::Execute if execute_service_manager
   action [:create, :start]
+end
+
+if node['docker']['service_manager'] == 'execute'
+  docker_service_execute('default', &service_def)
+else
+  docker_service('default', &service_def)
 end
